feat: Add endpoint for deleting activities by date in server.js
This commit adds a new endpoint in the server.js file to handle the deletion of activities based on a specific date. The endpoint is defined as '/activities/day/:date' and uses the Sequelize ORM to delete activities within the specified date range. If the deletion is successful, a 204 status code is returned along with the message 'Activities deleted'. If an error occurs, a 500 status code is returned.